+++ CHANGELOG.md
@@ -14,7 +14,7 @@ x.y.z Release notes (yyyy-MM-dd)
 ### Compatibility
 * Realm Studio: 15.0.0 or later.
 * APIs are backwards compatible with all previous releases in the 10.x.y series.
-* Carthage release for Swift is built with Xcode 15.4.0.
+* Carthage release for Swift is built with Xcode 1a.0.0.
 * CocoaPods: 1.10 or later.
 * Xcode: 15.3.0-16.1 beta.